Intermediate commit. Deviations still exist but much closer to expected output.

+++ src/main.cpp
@@ -117,9 +117,9 @@ int main()
 		  // Calculate and output the average weighted error of the particle filter over all time steps so far.
 		  vector<Particle> particles = pf.particles;
 		  int num_particles = particles.size();
-		  double highest_weight = -1.0;
+		  long double highest_weight = -1.0;
 		  Particle best_particle;
-		  double weight_sum = 0.0;
+		  long double weight_sum = 0.0;
 		  for (int i = 0; i < num_particles; ++i) {
 			if (particles[i].weight > highest_weight) {
 				highest_weight = particles[i].weight;
